+++ src/components/Chat/Chat.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./Chat.css";
-import ChatMessages from "../ChatMessages/ChatMessages";
+import { ChatMessages } from "../ChatMessages/ChatMessages";
 import { ChatList } from "../ChatList/ChatList";
 
 import { testData } from "./testData"; // Test data
