+++ src/components/Message/Message.tsx
@@ -1,12 +1,10 @@
 import React from "react";
-import "./ChatMessage.css";
+import "./Message.css";
 
 interface Props {
   message: string;
 }
 
-const Message = ({ message }: Props) => {
+export const Message = ({ message }: Props) => {
   return <li className="message">{message}</li>;
 };
-
-export default Message;
